Description.
-Changed the Home page to make a description of the project to the best
of my abilities.

-Changed the About page to add in kyra. Added a email
mathfun1000@gmail.com / PW: mathfunonethousand
this is just a temp email so I could get it to work on the about page,
but it works you click on it and it opens your default email up and you
can write an email to our team.

-Changed around tutorial and fill_in classes

-It took me some time to add resources in, I guess  you ahve to install
them on your machine if they are added to the project. Get Nuget as it
makes it easy to add these, all you do is just type in the name of it
and it installs it for you.

-I have an idea to put tutorial / fill in / unguided under a parent
class. I will write you guys an e-mail about this.