Graph.aspx displays graphs
I'll look into the graph layout options further but it looks like we may
be limited graphing positive numbers. If you have negative numbers to
being graphed it will graph them but the Y axis stays to the left
instead of being in the middle like students see in class.