Description...
-Changed Font to make it not look like a word doc. if you dont like it
please say something , or if you find a better font we should use. I
just didnt like the last one for the main site.

-Small UI changes to the orientation of the site and the boxes, Improved
the color to the site again. Yet again if you dont like it or think it
doesnt fit please say. I want it to look good.

-Multi problem is put in place and ready to be implemented. As of right
now I just have it filled out to make it ready to switch to java script.
The template i have now is just what it would look like as an idea. When
Daniel pushes his idea i will change it like the other problems to the
better looking JavaScript version