Mutli Hooked up, Changes to DB and Problems
Hooked up DB to do this i Had to change some things.

-Problems in DB
Had to make another column were it ID'd the type of problem. Daniel had
the Dynamic making his kind of problems. 1 is Default and its Daniels
problems, 2 is Graph, and 3 is Multi, Because these are all created
differently I thought It would be best to do this.

-Problems.aspx
I made it so it dynamicly makes mine and erics problems

-Eric
I was goign to change your code to connect to the DB, I have set it all
up for you all you have to do is put a connection in your code, and make
a problem in the database. If you want me to Do this just hit me up. I
have been learning and going over this for a bit.